feat: Rename Deposit to Receive and change maintab names Move/Sales
